refactor: add all the code in components
